[Refactor] DSV4 Async CAM forward && remove explicit register_afd() (#294)
<!-- markdownlint-disable -->
PLEASE FILL IN THE PR DESCRIPTION AND MAKE SURE THE CHECKLIST ITEMS HAVE
BEEN CONSIDERED.

## Purpose

This PR includes the following changes:

1. Isolate the DeepSeek V4 Async CAM forward path
- Move the model-owned two-stage Async CAM execution from
`deepseek_v4.py` into `deepseek_v4_async_cam_forward.py`.
- Keep the native vLLM-Ascend forward unchanged for requests without
Async CAM ubatch metadata.
- Preserve the existing Attention → CAM dispatch → remote FFN → CAM
combine schedule.

2. Improve the DSV4 forward integration
- Lazily load the Async CAM forward implementation only when the request
uses AFD Async MoE ubatching.
- Preserve the upstream forward signature and avoid copying the large
native DSV4 forward implementation.
- Add a regression test covering native-forward and Async-CAM
delegation.

3. Remove redundant AFD registration
- Remove the repeated `register_afd()` call from model configuration
handling.
- AFD registration remains handled by the installed plugin entrypoint.

## Issue

- Related issue(s): #
- Closing keyword, only if fully resolved: Closes #

## Scope

- In scope:
- Out of scope:

## Implementation Notes

<!--
Call out vLLM extension points, plugin-owned classes, compatibility
shims,
or any behavior that intentionally differs from the original AFD commit.
-->

## Test Plan

Cross-node AFD deployment:

- Attention: DP3 TP8
- FFN: DP8 EP8
- `HCCL_BUFFSIZE=4096`
- `max-num-batched-tokens=8192`

## Test Result

Request validation:

- Input length: 1024
- Output length: 128
- Concurrency: 2
- Requests: 4
- Successful: 4/4
- Failed: 0
- Output throughput: 9.80 tok/s
- Total token throughput: 88.18 tok/s
- Mean TTFT: 967.81 ms
- Mean TPOT: 198.08 ms

## Docs Impact

- Files updated:
- If none, reason:

---

<details>
<summary>Essential PR Checklist</summary>

- [ ] Purpose is clear and linked to public context when possible.
- [ ] Scope is bounded.
- [ ] Compatibility with vLLM v0.26.0 is considered.
- [ ] No changes are made to the vLLM source checkout.
- [ ] Plugin-owned classes or explicit dotted class paths are preferred
over monkey patches.
- [ ] Any compat shim or monkey patch is isolated, idempotent,
version-guarded, documented, and tested.
- [ ] Imports remain CPU-safe; CUDA-heavy work is delayed or GPU-gated.
- [ ] Validation evidence is included, including skipped GPU tests when
applicable.
- [ ] Documentation impact is stated.

</details>

---------
